not duplicating lines in resolv.conf

+++ src/OSes/windows_wsl2.py
@@ -36,6 +36,11 @@ def __generate_resolveconf():
         RESOLVCONF_DATA = f"{RESOLVCONF_HEADER}\n{RESOLVCONF_DATA}"
 
     resolv_script = f"""#!/usr/bin/env sh
+[ "$(ps a | grep tunnel | wc -l)" -le 1 ] && sudo {config.BASE_PATH}/bin/docker-dns tunnel
+
+if `grep -q \@docker-dns /etc/resolv.conf`; then
+    exit 0
+fi
 cp /etc/resolv.conf /tmp/resolv.ddns
 rm /etc/resolv.conf > /dev/null || true;
 cat <<EOL > /etc/resolv.conf
@@ -48,10 +53,7 @@ else
     cat /tmp/resolv.ddns >> /etc/resolv.conf
 fi
 rm /tmp/resolv.ddns
-
-[ "$(ps a | grep tunnel | wc -l)" -le 1 ] && sudo {config.BASE_PATH}/bin/docker-dns tunnel
 """
-    RESOLVCONF_DATA = f"{RESOLVCONF_HEADER}\n{RESOLVCONF_DATA}"
     open('/etc/resolv.conf', 'w').write(RESOLVCONF_DATA)
 
     open(f'{config.BASE_PATH}/bin/docker-dns.service.sh',
@@ -74,7 +76,7 @@ rm /tmp/resolv.ddns
 
     # Gotta find a better way to start that service, as real services does not work on WSL2 as you have microsoft's init.
     bashrc_content = open(f'{config.HOME}/.bashrc', 'r').read()
-    if 'TUNNEL_RUNNING' in bashrc_content:
+    if 'docker-dns end' in bashrc_content:
         bashrc_content_pre = bashrc_content.split('# docker-dns "service"')[0]
         bashrc_content_pos = bashrc_content.split('# docker-dns end')
         if len(bashrc_content_pos) == 1:
@@ -83,12 +85,11 @@ rm /tmp/resolv.ddns
             bashrc_content_pos = bashrc_content_pos[1]
         bashrc_content = f'{bashrc_content_pre}{bashrc_content_pos}'
 
-    service_script = f"""
-# docker-dns "service"  for windows wsl2
+    service_script = f"""# docker-dns "service"  for windows wsl2
 [ "$(ps a | grep tunnel | wc -l)" -le 1 ] && sudo {config.BASE_PATH}/bin/docker-dns.service.sh
 # docker-dns end
 """
-    bashrc_content = f"{bashrc_content}\n{service_script}"
+    bashrc_content = f"{bashrc_content}{service_script}"
     open(f'{config.HOME}/.bashrc', 'w').write(bashrc_content)
